// src/general/enums.ts
export enum Browser {
  EDGE = 'edge',
  CHROME = 'chrome',
  FIREFOX = 'firefox',
  OPERA = 'opera',
  SAFARI = 'safari',
  IE = 'ie',
  INTERNET_EXPLORER = 'explorer',
}

export enum Os {
  WIN = 'win',
  MAC = 'mac',
  LIN = 'lin',
  DROID = 'droid',
  IOS = 'ios',
}

export enum Device {
  MOBILE = 'mobile',
  TABLET = 'tablet',
  DESKTOP = 'desktop',
  MOB = 'mob',
  TAB = 'tab',
  DSK = 'dsk',
  TV = 'tv',
  BOT = 'bot',
}